introduce CashuPaymentRegistrar, with extracted logic from CashuPaymentService
