+++ Plugin/BTCPayServer.Plugins.Cashu/Services/CashuPaymentService.cs
@@ -2,9 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Security.Cryptography;
-using System.Text;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using BTCPayServer.Configuration;
@@ -50,7 +47,7 @@ public record CashuOperationContext(
 public class CashuPaymentService(
     StoreRepository storeRepository,
     InvoiceRepository invoiceRepository,
-    PaymentService paymentService,
+    CashuPaymentRegistrar cashuPaymentRegistrar,
     CashuPaymentMethodHandler handler,
     PaymentMethodHandlerDictionary handlers,
     LightningClientFactoryService lightningClientFactoryService,
@@ -278,7 +275,7 @@ public class CashuPaymentService(
                             ftx.MintUrl,
                             ProofState.Available
                         );
-                        await RegisterCashuPayment(ctx);
+                        await cashuPaymentRegistrar.Register(ctx);
                         return;
                     }
                 default:
@@ -318,7 +315,7 @@ public class CashuPaymentService(
             );
             //TODO: Pay partially or retry to recover missing proofs
         }
-        await RegisterCashuPayment(ctx);
+        await cashuPaymentRegistrar.Register(ctx);
     }
 
     /// <summary>
@@ -449,7 +446,7 @@ public class CashuPaymentService(
             var overpaidFeesReturned = (meltResponse.ChangeProofs?.Select(p => p.Amount).Sum() ?? 0L) * opCtx.UnitValue;
             var amountPaid = amountMelted + overpaidFeesReturned;
 
-            await RegisterCashuPayment(opCtx, amountPaid);
+            await cashuPaymentRegistrar.Register(opCtx, amountPaid);
 
             logs.PayServer.LogDebug(
                 "(Cashu) Melt success. Melted: {Melted} sat, fees returned: {FeesReturned} sat, total: {Total} sat",
@@ -519,7 +516,7 @@ public class CashuPaymentService(
                         }
                         var recoveredChange = (pollResult.ResultProofs?.Select(p => p.Amount).Sum() ?? 0L) * opCtx.UnitValue;
                         var recoveredAmount = (meltQuoteResponse.Invoice?.Amount ?? LightMoney.Zero) + recoveredChange;
-                        await RegisterCashuPayment(opCtx, recoveredAmount);
+                        await cashuPaymentRegistrar.Register(opCtx, recoveredAmount);
                         logs.PayServer.LogDebug(
                             "(Cashu) Melt recovered on retry for invoice {InvoiceId}. Amount: {Amount} sat",
                             opCtx.Invoice.Id,
@@ -586,7 +583,7 @@ public class CashuPaymentService(
 
             if (lnInvPaid)
             {
-                await RegisterCashuPayment(opCtx, meltQuoteResponse.Invoice?.Amount ?? LightMoney.Zero);
+                await cashuPaymentRegistrar.Register(opCtx, meltQuoteResponse.Invoice?.Amount ?? LightMoney.Zero);
 
                 logs.PayServer.LogDebug(
                     "(Cashu) Melt succeeded but SaveProofs failed. Invoice: {InvoiceId}. Error: {Error}",
@@ -649,146 +646,7 @@ public class CashuPaymentService(
         /// <summary>Input amount + currency combination can't produce a payment amount. Manual intervention required.</summary>
         UnresolvableAmount,
     }
-
-    /// <summary>
-    /// Registers payment for a recovered FailedTransaction. Idempotent.
-    /// Throws on transient errors (network, DB) so the caller keeps ftx unresolved and retries later.
-    /// Returns a terminal classification for permanent outcomes.
-    /// </summary>
-    public async Task<FtxPaymentRegistrationResult> RegisterPaymentForFailedTx(
-        FailedTransaction ftx,
-        CancellationToken ct = default
-    )
-    {
-        var invoice = await invoiceRepository.GetInvoice(ftx.InvoiceId, true);
-        if (invoice is null)
-            return FtxPaymentRegistrationResult.InvoiceMissing;
-
-        if (invoice.Status == InvoiceStatus.Settled)
-            return FtxPaymentRegistrationResult.Registered;
-
-        // GetTokenSatRate throws HttpRequestException on mint outage — propagate so ftx stays
-        // unresolved and the poller retries. Don't swallow as before (that hid lost payments).
-        var singleUnitPrice = await CashuUtils.GetTokenSatRate(
-            ftx.MintUrl,
-            ftx.Unit,
-            handler.Network.NBitcoinNetwork
-        );
-
-        var isOld = ftx is { InputAmount: 0, InputProofsJson: null };
-        LightMoney? paymentAmount = isOld switch
-        {
-            true when invoice.Currency is "SATS" => LightMoney.Satoshis(invoice.Price),
-            true when invoice.Currency is "BTC" => LightMoney.Coins(invoice.Price),
-            false => ftx.InputAmount * singleUnitPrice,
-            _ => null
-        };
-
-        if (paymentAmount is null)
-        {
-            logs.PayServer.LogError(
-                "(Cashu) Can't determine payment amount for failed tx {Id}: InputAmount={InputAmount} currency={Currency}. Manual registration required.",
-                ftx.Id, ftx.InputAmount, invoice.Currency
-            );
-            return FtxPaymentRegistrationResult.UnresolvableAmount;
-        }
-
-        await RegisterCashuPayment(invoice, paymentAmount, BuildPaymentIdForFtx(ftx));
-        return FtxPaymentRegistrationResult.Registered;
-    }
-
-    public Task RegisterCashuPayment(
-        CashuOperationContext ctx,
-        LightMoney value = null,
-        bool markPaid = true
-    ) => RegisterCashuPayment(
-        ctx.Invoice,
-        value ?? ctx.Value,
-        BuildPaymentId(ctx.Invoice.Id, ctx.Token.Proofs),
-        markPaid
-    );
-
-    /// <summary>
-    /// Registers a Cashu payment for the invoice. Idempotent: identified by deterministic paymentId
-    /// so retries, parallel callers, or partial-failure recoveries never double-count.
-    /// </summary>
-    public async Task RegisterCashuPayment(
-        InvoiceEntity invoice,
-        LightMoney value,
-        string paymentId,
-        bool markPaid = true
-    )
-    {
-        // Re-read invoice to avoid acting on stale payment list (caller's snapshot may miss a
-        // concurrent registration from another thread / manual retry / poller).
-        var fresh = await invoiceRepository.GetInvoice(invoice.Id, true) ?? invoice;
-        var alreadyRegistered = fresh
-            .GetPayments(false)
-            .Any(p => p.Id == paymentId);
-
-        if (!alreadyRegistered)
-        {
-            //set payment method fee to 0 so it won't be added to due for second time
-            var prompt = fresh.GetPaymentPrompt(CashuPlugin.CashuPmid);
-            if (prompt != null && prompt.PaymentMethodFee != 0.0m)
-            {
-                prompt.PaymentMethodFee = 0.0m;
-                await invoiceRepository.UpdatePrompt(fresh.Id, prompt);
-            }
-
-            var paymentData = new PaymentData
-            {
-                Id = paymentId,
-                Created = DateTimeOffset.UtcNow,
-                Status = PaymentStatus.Processing,
-                Currency = "BTC",
-                InvoiceDataId = fresh.Id,
-                Amount = value.ToDecimal(LightMoneyUnit.BTC),
-                PaymentMethodId = handler.PaymentMethodId.ToString(),
-            }.Set(fresh, handler, new CashuPaymentData());
-
-            // AddPayment swallows DbUpdateException on duplicate PK (Id, PaymentMethodId),
-            // returning null — so simultaneous callers are safe even if the pre-check missed.
-            await paymentService.AddPayment(paymentData);
-        }
-
-        if (markPaid && fresh.Status != InvoiceStatus.Settled)
-        {
-            await invoiceRepository.MarkInvoiceStatus(fresh.Id, InvoiceStatus.Settled);
-        }
-    }
-
-    private static readonly JsonSerializerOptions SecretJsonOptions = new()
-    {
-        Converters = { new SecretJsonConverter() }
-    };
-
-    /// <summary>
-    /// Deterministic payment identifier derived from invoice + input proof secrets.
-    /// Same (invoice, token) always yields the same ID so retries are idempotent at the DB layer.
-    /// </summary>
-    internal static string BuildPaymentId(string invoiceId, IEnumerable<Proof> inputProofs)
-    {
-        var secrets = inputProofs
-            .Select(p => JsonSerializer.Serialize(p.Secret, SecretJsonOptions))
-            .OrderBy(s => s, StringComparer.Ordinal);
-        var payload = $"{invoiceId}|{string.Join(",", secrets)}";
-        var hash = SHA256.HashData(Encoding.UTF8.GetBytes(payload));
-        return "cashu-" + Convert.ToHexString(hash).ToLowerInvariant();
-    }
-
-    /// <summary>
-    /// Deterministic payment identifier for a FailedTransaction. Prefers hash of InputProofs for
-    /// parity with the happy-path ID (so the same logical payment always maps to one record),
-    /// falls back to ftx.Id for legacy ftx rows missing InputProofsJson.
-    /// </summary>
-    internal static string BuildPaymentIdForFtx(FailedTransaction ftx)
-    {
-        var inputs = ftx.InputProofs;
-        return inputs.Length > 0
-            ? BuildPaymentId(ftx.InvoiceId, inputs)
-            : "cashu-ftx-" + ftx.Id.ToString("N");
-    }
+    
 
     private ILightningClient GetStoreLightningClient(StoreData store, BTCPayNetwork network)
     {
